split js into exports, lint files
