+++ src/scripts/main-dd-js.js
@@ -4,7 +4,6 @@ import './datadog-docs';
 
 import './components/announcement_banner';
 import './components/copy-code';
-import './components/codenav';
 import './components/global-modals';
 import './components/header';
 import './components/platforms';
