--- src/styles/components/_codenav.scss
@@ -1,25 +1,0 @@
-$ddpurple:    #632CA6;
-
-.codenav {
-  height:51px;
-  background:#000;
-  margin-bottom:20px;
-  padding-left:30px;
-  .nav-link {
-    color:#fff;
-    padding: 0.65rem 1rem;
-    -webkit-font-smoothing:subpixel-antialiased;
-    -moz-osx-font-smoothing:auto;
-    &.active,
-    &:hover {
-      border-bottom:1px solid $ddpurple;
-      color:$ddpurple;
-    }
-  }
-}
-
-@media (max-width: 991px) {
-  .codenav {
-    padding-left:0;
-  }
-}
