+++ src/styles/pages/_api.scss
@@ -328,8 +328,5 @@ $ddpurple: #632ca6;
             padding-top: 24px;
             padding-bottom: 24px;
         }
-        .codenav {
-            padding-left: 0;
-        }
     }
 }
