+++ src/styles/style.scss
@@ -34,7 +34,6 @@
 // Components
 @import "components/announcement_banner";
 @import "components/choice-select-box";
-@import "components/codenav";
 @import "components/footer";
 @import "components/global-modals";
 @import "components/header";
